+++ src/utils.js
@@ -1,7 +1,0 @@
-// Extend an object (a) with the values of another (b)
-module.exports.extend = (a, b) => {
-  Object.keys(b).forEach(key => {
-    a[key] = b[key]
-  })
-  return a
-}